move get_name_puzzle_conditions() into test utils. It's only used by tests
