+++ chia/full_node/mempool_check_conditions.py
@@ -4,18 +4,12 @@ import logging
 from typing import Optional
 
 from chia_rs import (
-    DONT_VALIDATE_SIGNATURE,
-    MEMPOOL_MODE,
-    G2Element,
     get_flags_for_height_and_constants,
-    run_block_generator,
-    run_block_generator2,
     run_chia_program,
 )
 from chia_rs import get_puzzle_and_solution_for_coin2 as get_puzzle_and_solution_for_coin_rust
 
 from chia.consensus.constants import ConsensusConstants
-from chia.consensus.cost_calculator import NPCResult
 from chia.types.blockchain_format.coin import Coin
 from chia.types.blockchain_format.program import Program
 from chia.types.blockchain_format.sized_bytes import bytes32
@@ -25,7 +19,7 @@ from chia.types.generator_types import BlockGenerator
 from chia.types.spend_bundle_conditions import SpendBundleConditions
 from chia.util.condition_tools import conditions_for_solution
 from chia.util.errors import Err
-from chia.util.ints import uint16, uint32, uint64
+from chia.util.ints import uint32, uint64
 from chia.wallet.puzzles.load_clvm import load_serialized_clvm_maybe_recompile
 
 DESERIALIZE_MOD = load_serialized_clvm_maybe_recompile(
@@ -33,38 +27,6 @@ DESERIALIZE_MOD = load_serialized_clvm_maybe_recompile(
 )
 
 log = logging.getLogger(__name__)
-
-
-def get_name_puzzle_conditions(
-    generator: BlockGenerator,
-    max_cost: int,
-    *,
-    mempool_mode: bool,
-    height: uint32,
-    constants: ConsensusConstants,
-) -> NPCResult:
-    flags = get_flags_for_height_and_constants(height, constants) | DONT_VALIDATE_SIGNATURE
-
-    if mempool_mode:
-        flags |= MEMPOOL_MODE
-
-    if height >= constants.HARD_FORK_HEIGHT:
-        run_block = run_block_generator2
-    else:
-        run_block = run_block_generator
-
-    try:
-        block_args = generator.generator_refs
-        err, result = run_block(bytes(generator.program), block_args, max_cost, flags, G2Element(), None, constants)
-        assert (err is None) != (result is None)
-        if err is not None:
-            return NPCResult(uint16(err), None)
-        else:
-            assert result is not None
-            return NPCResult(None, result)
-    except BaseException:
-        log.exception("get_name_puzzle_condition failed")
-        return NPCResult(uint16(Err.GENERATOR_RUNTIME_ERROR.value), None)
 
 
 def get_puzzle_and_solution_for_coin(
